+++ spine-unity/Assets/Spine/Runtime/spine-unity/Components/Base/SkeletonAnimationBase.cs
@@ -62,10 +62,22 @@ namespace Spine.Unity {
 		ISkeletonAnimation, ISkeletonComponent, IHasSkeletonRenderer,
 		ISkeletonRendererEvents, IHasModifyableSkeletonDataAsset, IUpgradable {
 
+		public delegate void DeltaTimeOverrideDelegate (SkeletonAnimationBase animation, ref float deltaTime);
+
 		[SerializeField] protected UpdateTiming updateTiming = UpdateTiming.InUpdate;
 		protected int frameOfLastUpdate = -1;
 		protected ISkeletonRenderer skeletonRenderer;
 		protected bool skipUpdate = false;
+		protected float resolvedDeltaTime = 0f;
+		protected DeltaTimeOverrideDelegate deltaTimeOverride;
+
+		/// <summary>Allows advancing by custom delta time for e.g. stepped animation updates in a threaded context.
+		/// The callback may change <see cref="UpdateMode"/> to disable or limit skeleton updates for certain frames.
+		/// Called on the main thread once before animation update.</summary>
+		public DeltaTimeOverrideDelegate DeltaTimeOverride {
+			get { return deltaTimeOverride; }
+			set { deltaTimeOverride = value; }
+		}
 
 		public UpdateTiming UpdateTiming {
 			get { return updateTiming; }
@@ -408,8 +420,9 @@ namespace Spine.Unity {
 		/// <summary>Calls <see cref="Update()"/> if it has not yet been called this frame.</summary>
 		public virtual void UpdateOncePerFrame (float deltaTime) {
 			if (frameOfLastUpdate != Time.frameCount) {
+				resolvedDeltaTime = deltaTime;
 				MainThreadBeforeUpdateInternal();
-				UpdateInternal(deltaTime, Time.frameCount, calledFromOnlyMainThread: true);
+				UpdateInternal(resolvedDeltaTime, Time.frameCount, calledFromOnlyMainThread: true);
 			}
 		}
 
@@ -418,6 +431,13 @@ namespace Spine.Unity {
 		/// To be followed by a potentially threaded call to <see cref="UpdateInternal"/>.
 		/// </summary>
 		public virtual void MainThreadBeforeUpdateInternal () {
+#if USE_THREADED_ANIMATION_UPDATE
+			if (isUpdatedExternally)
+				resolvedDeltaTime = UsedExternalDeltaTime;
+#endif
+			if (deltaTimeOverride != null)
+				deltaTimeOverride(this, ref resolvedDeltaTime);
+
 			if (skeletonRenderer == null || !skeletonRenderer.IsValid || skeletonRenderer.Freeze || !this.IsValid
 				|| skeletonRenderer.UpdateMode < UpdateMode.OnlyAnimationStatus) {
 				skipUpdate = true;
@@ -433,7 +453,7 @@ namespace Spine.Unity {
 
 #if USE_THREADED_ANIMATION_UPDATE
 		public virtual void UpdateExternal (int currentFrameCount, bool calledFromOnlyMainThread = true) {
-			UpdateInternal(UsedExternalDeltaTime, currentFrameCount, calledFromOnlyMainThread);
+			UpdateInternal(resolvedDeltaTime, currentFrameCount, calledFromOnlyMainThread);
 		}
 #endif
 		public virtual void UpdateInternal (float deltaTime, int currentFrameCount, bool calledFromOnlyMainThread = true) {
@@ -466,7 +486,7 @@ namespace Spine.Unity {
 				if (skipUpdate)
 					return CoroutineIterator.Done;
 				frameOfLastUpdate = currentFrameCount;
-				UpdateAnimationStatus(UsedExternalDeltaTime);
+				UpdateAnimationStatus(resolvedDeltaTime);
 				skeletonRenderer.ApplyTransformMovementToPhysics();
 
 				if (skeletonRenderer.UpdateMode == UpdateMode.OnlyAnimationStatus)
@@ -487,8 +507,9 @@ namespace Spine.Unity {
 		/// <summary>Progresses the AnimationState according to the given deltaTime, and applies it to the Skeleton.
 		/// Use Time.deltaTime to update manually. Use deltaTime 0 to update without progressing the time.</summary>
 		public virtual void Update (float deltaTime) {
+			resolvedDeltaTime = deltaTime;
 			MainThreadBeforeUpdateInternal();
-			UpdateInternal(deltaTime, Time.frameCount, calledFromOnlyMainThread: true);
+			UpdateInternal(resolvedDeltaTime, Time.frameCount, calledFromOnlyMainThread: true);
 		}
 
 		public virtual void ApplyAnimation (bool calledFromMainThread = true) {